Migration and refactoring, process pending (3)
- Add purchase related technical module
- Add po2so scenario
