+++ README.md
@@ -1,7 +1,7 @@
 Odoo2Odoo Connector Modules
 ============================
 
-*Odoo2Odoo Connector enables a bi-directional communication between two (or more), sepearte Odoo instances.*
+*Odoo2Odoo Connector enables a bi-directional communication between two (or more), separate Odoo instances.*
 
 The Odoo2Odoo Connector is based on the [OpenERP Connector framework](https://github.com/OCA/connector).
