+++ odooconnector_po2so/__openerp__.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+# © 2015 Malte Jacobi (maljac @ github)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+{
+    'name': 'Odoo Connector - PO2SO',
+    'summary': 'PO2SO Process Addon',
+    'version': '0.0.1',
+    'category': 'Connector',
+    'website': 'https://odoo-community.org/',
+    'author': 'IBO Institut (Malte Jacobi), Odoo Community Association (OCA)',
+    'license': 'AGPL-3',
+    'application': True,
+    'installable': True,
+    'external_dependencies': {
+    },
+    'depends': [
+        'base',
+        'purchase',
+        'odooconnector_base',
+        'odooconnector_purchase'
+    ],
+    'data': [
+    ],
+    'demo': [
+    ],
+}
